+++ api/src/test/java/org/pmiops/workbench/access/AccessSyncServiceTest.java
@@ -1,6 +1,7 @@
 package org.pmiops.workbench.access;
 
 import static com.google.common.truth.Truth.assertThat;
+import static com.google.common.truth.Truth.assertWithMessage;
 import static com.google.common.truth.Truth8.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.Mockito.verifyZeroInteractions;
@@ -15,6 +16,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
+import java.util.function.Supplier;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.pmiops.workbench.FakeClockConfiguration;
@@ -32,6 +34,7 @@ import org.pmiops.workbench.db.model.DbUser;
 import org.pmiops.workbench.db.model.DbUserAccessModule;
 import org.pmiops.workbench.exceptions.NotFoundException;
 import org.pmiops.workbench.firecloud.FireCloudService;
+import org.pmiops.workbench.firecloud.model.FirecloudNihStatus;
 import org.pmiops.workbench.google.DirectoryService;
 import org.pmiops.workbench.institution.InstitutionService;
 import org.pmiops.workbench.moodle.ApiException;
@@ -67,6 +70,7 @@ public class AccessSyncServiceTest {
 
   @MockBean private ComplianceService mockComplianceService;
   @MockBean private DirectoryService mockDirectoryService;
+  @MockBean private FireCloudService mockFireCloudService;
 
   private static WorkbenchConfig providedWorkbenchConfig;
   private static List<DbAccessModule> providedAccessModules;
@@ -336,6 +340,79 @@ public class AccessSyncServiceTest {
     accessSyncService.syncDuccVersionStatus(user, Agent.asSystem());
 
     assertModuleCompletionMissing(DbAccessModuleName.DATA_USER_CODE_OF_CONDUCT, user);
+  }
+
+  @Test
+  public void testSyncEraCommonsStatus() {
+    FirecloudNihStatus nihStatus = new FirecloudNihStatus();
+    nihStatus.setLinkedNihUsername("nih-user");
+    // FireCloud stores the NIH status in seconds, not msecs.
+    final long FC_LINK_EXPIRATION_SECONDS = START_INSTANT.toEpochMilli() / 1000;
+    nihStatus.setLinkExpireTime(FC_LINK_EXPIRATION_SECONDS);
+
+    when(mockFireCloudService.getNihStatus()).thenReturn(nihStatus);
+
+    accessSyncService.syncEraCommonsStatus();
+
+    DbUser user = userDao.findUserByUsername(USERNAME);
+    assertModuleCompletionEqual(
+        DbAccessModuleName.ERA_COMMONS, user, Timestamp.from(START_INSTANT));
+
+    assertThat(user.getEraCommonsLinkExpireTime()).isEqualTo(Timestamp.from(START_INSTANT));
+    assertThat(user.getEraCommonsLinkedNihUsername()).isEqualTo("nih-user");
+
+    // Completion timestamp should not change when the method is called again.
+    tick();
+    accessSyncService.syncEraCommonsStatus();
+
+    assertModuleCompletionEqual(
+        DbAccessModuleName.ERA_COMMONS, user, Timestamp.from(START_INSTANT));
+  }
+
+  @Test
+  public void testClearsEraCommonsStatus() {
+    // Put the test user in a state where eRA commons is completed.
+    DbUser testUser = userDao.findUserByUsername(USERNAME);
+    testUser.setEraCommonsLinkedNihUsername("nih-user");
+    testUser = userDao.save(testUser);
+
+    accessModuleService.updateCompletionTime(
+        testUser, DbAccessModuleName.ERA_COMMONS, Timestamp.from(START_INSTANT));
+
+    accessSyncService.syncEraCommonsStatus();
+
+    DbUser retrievedUser = userDao.findUserByUsername(USERNAME);
+    assertModuleCompletionEqual(DbAccessModuleName.ERA_COMMONS, retrievedUser, null);
+  }
+
+  @Test
+  public void testSyncEraCommonsStatus_lastModified() {
+    // User starts without eRA commons.
+    Supplier<Timestamp> getLastModified =
+        () -> userDao.findUserByUsername(USERNAME).getLastModifiedTime();
+    Timestamp modifiedTime0 = getLastModified.get();
+
+    when(mockFireCloudService.getNihStatus())
+        .thenReturn(
+            new FirecloudNihStatus()
+                .linkedNihUsername("nih-user")
+                // FireCloud stores the NIH status in seconds, not msecs.
+                .linkExpireTime(START_INSTANT.toEpochMilli() / 1000));
+
+    tick();
+    accessSyncService.syncEraCommonsStatus();
+    Timestamp modifiedTime1 = getLastModified.get();
+    assertWithMessage(
+            "modified time should change when eRA commons status changes, want %s < %s",
+            modifiedTime0, modifiedTime1)
+        .that(modifiedTime0.before(modifiedTime1))
+        .isTrue();
+
+    accessSyncService.syncEraCommonsStatus();
+    assertWithMessage(
+            "modified time should not change on sync, if eRA commons status doesn't change")
+        .that(modifiedTime1)
+        .isEqualTo(getLastModified.get());
   }
 
   private Optional<Timestamp> getModuleCompletionTime(DbAccessModuleName moduleName, DbUser user) {
